Fixing after pull request
