+++ src/main/java/com/es/phoneshop/dao/Dao.java
@@ -3,5 +3,5 @@ package com.es.phoneshop.dao;
 public interface Dao<T> {
     T getItem(Long id);
 
-    void save(T o);
+    void save(T item);
 }
