--- src/main/java/com/es/phoneshop/exception/PastDateException.java
@@ -1,4 +1,0 @@
-package com.es.phoneshop.exception;
-
-public class PastDateException extends Exception{
-}
