+++ src/main/java/com/es/phoneshop/web/servlet/CartPageServlet.java
@@ -1,13 +1,9 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
-import com.es.phoneshop.dao.impl.ArrayListProductDao;
-import com.es.phoneshop.dao.ProductDao;
 import com.es.phoneshop.exception.OutOfQuantityException;
 import com.es.phoneshop.exception.OutOfStockException;
 import com.es.phoneshop.exception.ParseToIntegerException;
 import com.es.phoneshop.model.cart.Cart;
-import com.es.phoneshop.model.sort.SortField;
-import com.es.phoneshop.model.sort.SortOrder;
 import com.es.phoneshop.service.CartService;
 import com.es.phoneshop.service.impl.HttpSessionCartService;
 import com.es.phoneshop.utils.CartLoader;
@@ -24,10 +20,8 @@ import java.io.IOException;
 import java.text.ParseException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
-public class ProductListPageServlet extends HttpServlet {
-    private ProductDao productDao;
+public class CartPageServlet extends HttpServlet {
     private CartService cartService;
     private ParserService parserService;
     private ErrorService errorService;
@@ -35,43 +29,34 @@ public class ProductListPageServlet extends HttpServlet {
     @Override
     public void init(ServletConfig config) throws ServletException {
         super.init(config);
-        productDao = ArrayListProductDao.getInstance();
         cartService = HttpSessionCartService.getInstance();
         parserService = ParserService.getInstance();
         errorService = ErrorService.getInstance();
     }
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String query = request.getParameter("query");
-        String sortField = request.getParameter("sort");
-        String sortOrder = request.getParameter("order");
-        request.setAttribute("products", productDao.findProducts(query,
-                Optional.ofNullable(sortField).map(SortField::valueOf).orElse(null),
-                Optional.ofNullable(sortOrder).map(SortOrder::valueOf).orElse(null)));
-        request.getRequestDispatcher("/WEB-INF/pages/main/productList.jsp").forward(request, response);
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher("/WEB-INF/pages/main/cart.jsp").forward(request, response);
     }
 
     @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String[] productIds = request.getParameterValues("productId");
         String[] quantities = request.getParameterValues("quantity");
         Map<Long, String> errors = new HashMap<>();
         CartLoader cartLoader = new HttpSessionCartLoader(request);
         Cart cart = cartService.getCart(cartLoader);
         for (int i = 0; i < productIds.length; i++) {
-            if (request.getParameter("button" + productIds[i]) != null) {
-                Long productId = Long.valueOf(productIds[i]);
-                try {
-                    int quantity = parserService.parseQuantity(quantities[i], request);
-                    cartService.add(cart, productId, quantity);
-                } catch (ParseException | ParseToIntegerException | OutOfQuantityException | OutOfStockException e) {
-                    errorService.handleErrors(errors, productId, e);
-                }
+            Long productId = Long.valueOf(productIds[i]);
+            try {
+                int quantity = parserService.parseQuantity(quantities[i], request);
+                cartService.update(cart, productId, quantity);
+            } catch (ParseException | ParseToIntegerException | OutOfQuantityException | OutOfStockException e) {
+                errorService.handleErrors(errors, productId, e);
             }
         }
         if (errors.isEmpty()) {
-            response.sendRedirect(request.getContextPath() + "/products?message=Added to cart successfully");
+            response.sendRedirect(request.getContextPath() + "/cart?message=Cart updated successfully");
         } else {
             request.setAttribute("errors", errors);
             doGet(request, response);
