+++ src/main/java/com/es/phoneshop/web/servlet/ProductDetailsPageServlet.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
 import com.es.phoneshop.dao.ProductDao;
 import com.es.phoneshop.dao.impl.ArrayListProductDao;
@@ -43,7 +43,7 @@ public class ProductDetailsPageServlet extends HttpServlet {
     }
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         RecentlyViewedProductsLoader viewedProductsLoader = new HttpSessionRecentlyViewedProductsLoader(request);
         Long productId = parserService.parseProductId(request);
         request.setAttribute("product", productDao.getItem(productId));
