+++ src/main/java/com/es/phoneshop/web/servlet/ProductListPageServlet.java
@@ -1,9 +1,13 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
+import com.es.phoneshop.dao.impl.ArrayListProductDao;
+import com.es.phoneshop.dao.ProductDao;
 import com.es.phoneshop.exception.OutOfQuantityException;
 import com.es.phoneshop.exception.OutOfStockException;
 import com.es.phoneshop.exception.ParseToIntegerException;
 import com.es.phoneshop.model.cart.Cart;
+import com.es.phoneshop.model.sort.SortField;
+import com.es.phoneshop.model.sort.SortOrder;
 import com.es.phoneshop.service.CartService;
 import com.es.phoneshop.service.impl.HttpSessionCartService;
 import com.es.phoneshop.utils.CartLoader;
@@ -20,8 +24,10 @@ import java.io.IOException;
 import java.text.ParseException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
-public class CartPageServlet extends HttpServlet {
+public class ProductListPageServlet extends HttpServlet {
+    private ProductDao productDao;
     private CartService cartService;
     private ParserService parserService;
     private ErrorService errorService;
@@ -29,14 +35,21 @@ public class CartPageServlet extends HttpServlet {
     @Override
     public void init(ServletConfig config) throws ServletException {
         super.init(config);
+        productDao = ArrayListProductDao.getInstance();
         cartService = HttpSessionCartService.getInstance();
         parserService = ParserService.getInstance();
         errorService = ErrorService.getInstance();
     }
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.getRequestDispatcher("/WEB-INF/pages/main/cart.jsp").forward(request, response);
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String query = request.getParameter("query");
+        String sortField = request.getParameter("sort");
+        String sortOrder = request.getParameter("order");
+        request.setAttribute("products", productDao.findProducts(query,
+                Optional.ofNullable(sortField).map(SortField::valueOf).orElse(null),
+                Optional.ofNullable(sortOrder).map(SortOrder::valueOf).orElse(null)));
+        request.getRequestDispatcher("/WEB-INF/pages/main/productList.jsp").forward(request, response);
     }
 
     @Override
@@ -47,16 +60,18 @@ public class CartPageServlet extends HttpServlet {
         CartLoader cartLoader = new HttpSessionCartLoader(request);
         Cart cart = cartService.getCart(cartLoader);
         for (int i = 0; i < productIds.length; i++) {
-            Long productId = Long.valueOf(productIds[i]);
-            try {
-                int quantity = parserService.parseQuantity(quantities[i], request);
-                cartService.update(cart, productId, quantity);
-            } catch (ParseException | ParseToIntegerException | OutOfQuantityException | OutOfStockException e) {
-                errorService.handleErrors(errors, productId, e);
+            if (request.getParameter("button" + productIds[i]) != null) {
+                Long productId = Long.valueOf(productIds[i]);
+                try {
+                    int quantity = parserService.parseQuantity(quantities[i], request);
+                    cartService.add(cart, productId, quantity);
+                } catch (ParseException | ParseToIntegerException | OutOfQuantityException | OutOfStockException e) {
+                    errorService.handleErrors(errors, productId, e);
+                }
             }
         }
         if (errors.isEmpty()) {
-            response.sendRedirect(request.getContextPath() + "/cart?message=Cart updated successfully");
+            response.sendRedirect(request.getContextPath() + "/products?message=Added to cart successfully");
         } else {
             request.setAttribute("errors", errors);
             doGet(request, response);
