+++ src/main/java/com/es/phoneshop/web/servlet/ProductPriceHistoryPageServlet.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
 import com.es.phoneshop.dao.ProductDao;
 import com.es.phoneshop.dao.impl.ArrayListProductDao;
@@ -25,7 +25,7 @@ public class ProductPriceHistoryPageServlet extends HttpServlet {
     }
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String productId = request.getPathInfo().substring(1);
         Product product = productDao.getItem(Long.valueOf(productId));
         request.setAttribute("product", product);
