+++ src/main/webapp/WEB-INF/pages/exception/errorOrderNotFound.jsp
@@ -3,7 +3,7 @@
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
 <%@ taglib prefix="tags" tagdir="/WEB-INF/tags" %>
 
-<tags:master pageTitle="Product not found">
+<tags:master pageTitle="Order not found">
   <h1>
     Sorry, your order is not found
   </h1>
