+++ src/main/webapp/WEB-INF/web.xml
@@ -11,12 +11,12 @@
   </context-param>
 
   <listener>
-    <listener-class>com.es.phoneshop.model.product.ProductDemodataServletContextListener</listener-class>
+    <listener-class>com.es.phoneshop.web.listener.ProductDemodataServletContextListener</listener-class>
   </listener>
 
   <servlet>
     <servlet-name>productList</servlet-name>
-    <servlet-class>com.es.phoneshop.web.ProductListPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.ProductListPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -26,7 +26,7 @@
 
   <servlet>
     <servlet-name>product</servlet-name>
-    <servlet-class>com.es.phoneshop.web.ProductDetailsPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.ProductDetailsPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -36,7 +36,7 @@
 
   <servlet>
     <servlet-name>history</servlet-name>
-    <servlet-class>com.es.phoneshop.web.ProductPriceHistoryPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.ProductPriceHistoryPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -46,7 +46,7 @@
 
   <servlet>
     <servlet-name>cart</servlet-name>
-    <servlet-class>com.es.phoneshop.web.CartPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.CartPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -56,7 +56,7 @@
 
   <servlet>
     <servlet-name>deleteCartItem</servlet-name>
-    <servlet-class>com.es.phoneshop.web.CartItemDeleteServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.CartItemDeleteServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -66,7 +66,7 @@
 
   <servlet>
     <servlet-name>minicart</servlet-name>
-    <servlet-class>com.es.phoneshop.web.MiniCartServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.MiniCartServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -76,7 +76,7 @@
 
   <servlet>
     <servlet-name>checkout</servlet-name>
-    <servlet-class>com.es.phoneshop.web.CheckoutPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.CheckoutPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -86,7 +86,7 @@
 
   <servlet>
     <servlet-name>orderOverview</servlet-name>
-    <servlet-class>com.es.phoneshop.web.OrderOverviewPageServlet</servlet-class>
+    <servlet-class>com.es.phoneshop.web.servlet.OrderOverviewPageServlet</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -116,7 +116,7 @@
 
   <filter>
     <filter-name>dosFilter</filter-name>
-    <filter-class>com.es.phoneshop.web.DosFilter</filter-class>
+    <filter-class>com.es.phoneshop.web.filter.DosFilter</filter-class>
   </filter>
 
   <filter-mapping>
