+++ src/test/java/com/es/phoneshop/web/listener/ProductDemodataServletContextListenerTest.java
@@ -1,6 +1,7 @@
-package com.es.phoneshop.model.product;
+package com.es.phoneshop.web.listener;
 
 import com.es.phoneshop.dao.impl.ArrayListProductDao;
+import com.es.phoneshop.web.listener.ProductDemodataServletContextListener;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
