+++ src/test/java/com/es/phoneshop/web/servlet/CheckoutPageServletTest.java
@@ -1,10 +1,15 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
-import com.es.phoneshop.dao.impl.ArrayListProductDao;
+import com.es.phoneshop.dao.OrderDao;
+import com.es.phoneshop.dao.impl.ArrayListOrderDao;
 import com.es.phoneshop.model.cart.Cart;
 import com.es.phoneshop.model.cart.CartItem;
+import com.es.phoneshop.model.order.Order;
+import com.es.phoneshop.model.order.PaymentMethod;
 import com.es.phoneshop.model.product.Product;
+import com.es.phoneshop.service.OrderService;
 import com.es.phoneshop.service.impl.HttpSessionCartService;
+import com.es.phoneshop.service.impl.HttpSessionOrderService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,16 +26,13 @@ import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.Currency;
-import java.util.Locale;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CartPageServletTest {
-    @Mock
-    private ArrayListProductDao productDao;
+public class CheckoutPageServletTest {
     @Mock
     private HttpServletRequest request;
     @Mock
@@ -39,36 +41,35 @@ public class CartPageServletTest {
     private RequestDispatcher requestDispatcher;
     @Mock
     private HttpSession session;
-    private Product product1;
-    private Product product2;
     private Cart cart;
-    private String[] productId = new String[]{"1", "2"};
-    private String[] stringQuantity = new String[]{"5", "2"};
+    private Order order;
+    private OrderService orderService;
+    private OrderDao orderDao;
     private ServletConfig config;
 
     @InjectMocks
-    private CartPageServlet servlet = new CartPageServlet();
+    private CheckoutPageServlet servlet = new CheckoutPageServlet();
 
-    public CartPageServletTest() {
+    public CheckoutPageServletTest() {
     }
 
     @Before
     public void setup() throws ServletException {
-        productDao = ArrayListProductDao.getInstance();
+        orderDao = ArrayListOrderDao.getInstance();
+        orderService = HttpSessionOrderService.getInstance();
         when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
         when(request.getSession()).thenReturn(session);
         cart = new Cart();
+        order = new Order();
         Currency usd = Currency.getInstance("USD");
-        product1 = new Product(1L, "test-1", "Samsung Galaxy S", new BigDecimal(300), usd, 100, "https://raw.githubusercontent.com/andrewosipenko/phoneshop-ext-images/master/manufacturer/Samsung/Samsung%20Galaxy%20S.jpg");
-        product2 = new Product(2L, "test-2", "Samsung Galaxy S II", new BigDecimal(200), usd, 10, "https://raw.githubusercontent.com/andrewosipenko/phoneshop-ext-images/master/manufacturer/Samsung/Samsung%20Galaxy%20S%20II.jpg");
+        Product product1 = new Product(1L, "test-1", "Samsung Galaxy S", new BigDecimal(300), usd, 100, "https://raw.githubusercontent.com/andrewosipenko/phoneshop-ext-images/master/manufacturer/Samsung/Samsung%20Galaxy%20S.jpg");
+        Product product2 = new Product(2L, "test-2", "Samsung Galaxy S II", new BigDecimal(200), usd, 10, "https://raw.githubusercontent.com/andrewosipenko/phoneshop-ext-images/master/manufacturer/Samsung/Samsung%20Galaxy%20S%20II.jpg");
         cart.getItems().add(new CartItem(product1, 2));
         cart.getItems().add(new CartItem(product2, 3));
+        cart.setTotalCost(BigDecimal.valueOf(1200));
         when(request.getSession().getAttribute(HttpSessionCartService.class.getName() + ".cart")).thenReturn(cart);
-        when(request.getLocale()).thenReturn(Locale.ENGLISH);
-        when(request.getParameterValues("quantity")).thenReturn(stringQuantity);
-        when(request.getParameterValues("productId")).thenReturn(productId);
-        productDao.save(product1);
-        productDao.save(product2);
+        order = orderService.getOrder(cart);
+        orderDao.save(order);
         servlet.init(config);
     }
 
@@ -77,12 +78,19 @@ public class CartPageServletTest {
         servlet.doGet(request, response);
 
         verify(requestDispatcher).forward(request, response);
+        request.setAttribute("paymentMethods", PaymentMethod.values());
     }
 
     @Test
     public void testDoPost() throws ServletException, IOException {
+        when(request.getParameter("firstName")).thenReturn("Aaa");
+        when(request.getParameter("lastName")).thenReturn("Bbb");
+        when(request.getParameter("deliveryDate")).thenReturn("2021-03-03");
+        when(request.getParameter("deliveryAddress")).thenReturn("Ccc");
+        when(request.getParameter("phone")).thenReturn("+375293632323");
+        when(request.getParameter("payment")).thenReturn("CACHE");
         servlet.doPost(request, response);
 
-        verify(response).sendRedirect(request.getContextPath() + "/cart?message=Cart updated successfully");
+        verify(response).sendRedirect(anyString());
     }
 }
