+++ src/test/java/com/es/phoneshop/web/servlet/MiniCarteServletTest.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
 import com.es.phoneshop.model.cart.Cart;
 import com.es.phoneshop.service.impl.HttpSessionCartService;
